% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeadSorted.Saliva.EPIC.compTable.R
\docType{data}
\name{BeadSorted.Saliva.EPIC.compTable}
\alias{BeadSorted.Saliva.EPIC.compTable}
\title{Cell Composition Association Table}
\format{Name: CpG identifier from Illumina EPIC \n
other stuff}
\description{
Association of each probe in the Illumina EPIC with saliva cell composition.
}
