% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeadSorted.Saliva.EPIC.compTable.R
\docType{data}
\name{BeadSorted.Saliva.EPIC.compTable}
\alias{BeadSorted.Saliva.EPIC.compTable}
\title{Cell Proportion Association Table}
\format{
A data frame with 795694 rows and 7 columns
\describe{
\item{probeName}{CpG identifier}
\item{t-statistic}{test statistic for the t-test between immune and epithelial compositions}
\item{p-value}{p-value for the t-test between immune and epithelial compositions}
\item{averageMethylationImmune}{average methylation beta value across immune cell samples}
\item{averageMethylationEpithelial}{average methylation beta value across epithelial cell samples}
\item{low}{minimum methylation beta across all samples}
\item{high}{maximum methylation beta across all samples}
}
}
\usage{
BeadSorted.Saliva.EPIC.compTable
}
\description{
Association of each probe in the Illumina EPIC array with saliva cell type composition.
}
\examples{
library(ExperimentHub)
query(ExperimentHub(), "BeadSorted.Saliva.EPIC.compTable")

BeadSorted.Saliva.EPIC.compTable <- ExperimentHub()[["OURID"]]
BeadSorted.Saliva.EPIC.compTable

}
\keyword{datasets}
