% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeadSorted.Saliva.EPIC.estimates.R
\docType{data}
\name{BeadSorted.Saliva.EPIC.estimates}
\alias{BeadSorted.Saliva.EPIC.estimates}
\title{Estimated cell composition for samples}
\format{A data frame with 60 rows and 2 variables
\describe{
\item{sampid}{Sample identification number}
\item{immuneCells}{Estimated proportion of immune cells in sample}
\item{epithelialCells}{Estimate proportion of epithelial cells in sample}
}}
\usage{
BeadSorted.Saliva.EPIC.estimates
}
\description{
Estimated cell composition for samples
}
\keyword{datasets}
